kotlin: split LogLevel into its own file
Cleaner/easier to find, as well as maintaining parity with iOS.
